requirements.txt
absl-py==1.4.0
adlfs==2024.2.0
aiobotocore==2.11.2
aiohttp==3.9.3
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiooss2==0.2.9
aiosignal==1.3.1
alembic==1.13.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
apache-airflow==2.8.1
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-common-sql==1.11.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.9.1
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.7.1
apache-beam==2.53.0
apispec==6.4.0
appdirs==1.4.4
argcomplete==3.2.2
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.0
attrs==23.2.0
azure-core==1.30.0
azure-datalake-store==0.0.53
azure-identity==1.15.0
azure-storage-blob==12.19.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boto3==1.34.34
botocore==1.34.34
cachelib==0.9.0
cachetools==5.3.2
catalogue==2.0.10
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpathlib==0.16.0
cloudpickle==2.2.1
cmake==3.28.1
colorama==0.4.6
colorlog==4.8.0
confection==0.1.4
configobj==5.0.8
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.0
crcmod==1.7
cron-descriptor==1.4.3
croniter==2.0.1
cryptography==42.0.4
cycler==0.12.1
cymem==2.0.8
databricks-cli==0.18.0
decorator==5.1.1
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.1.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.5.0
docker==7.0.0
docopt==0.6.2
docutils==0.20.1
dpath==2.1.6
dulwich==0.21.7
dvc==3.48.0
dvc-azure==3.1.0
dvc-data==3.13.0
dvc-gdrive==3.0.1
dvc-gs==3.0.1
dvc-hdfs==3.0.0
dvc-http==2.32.0
dvc-objects==5.0.0
dvc-oss==3.0.0
dvc-render==1.0.1
dvc-s3==3.0.1
dvc-ssh==4.1.1
dvc-studio-client==0.20.0
dvc-task==0.3.0
dvc-webdav==3.0.0
dvc-webhdfs==3.1.0
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.2.0
fastapi==0.109.2
fastavro==1.9.3
fasteners==0.19
filelock==3.13.1
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
funcy==2.0
gast==0.5.4
gcsfs==2024.2.0
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.17.0
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==2.27.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.41.0
google-cloud-bigquery==3.17.2
google-cloud-bigquery-storage==2.24.0
google-cloud-bigtable==2.23.0
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-dlp==3.15.1
google-cloud-language==2.13.1
google-cloud-pubsub==2.19.4
google-cloud-pubsublite==1.9.0
google-cloud-recommendations-ai==0.10.8
google-cloud-resource-manager==1.12.1
google-cloud-spanner==3.42.0
google-cloud-storage==2.15.0
google-cloud-videointelligence==2.13.1
google-cloud-vision==3.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-re2==1.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grandalf==0.8
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpc-interceptor==0.15.4
grpcio==1.60.1
grpcio-status==1.48.2
gto==1.7.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
hdfs==2.7.3
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.21.4
hydra-core==1.3.2
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
isodate==0.6.1
iterative-telemetry==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
Js2Py==0.74
jsonpickle==3.0.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaggle==1.6.6
keras==2.15.0
kiwisolver==1.4.5
knack==0.11.0
kneed==0.8.5
kombu==5.3.5
langcodes==3.3.0
lazy-object-proxy==1.10.0
libclang==16.0.6
limits==3.9.0
linkify-it-py==2.0.3
lit==17.0.6
lockfile==0.12.2
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.2.0
mlflow==2.10.1
more-itertools==10.2.0
mpmath==1.3.0
msal==1.27.0
msal-extensions==1.1.0
multidict==6.0.5
munkres==1.1.4
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
omegaconf==2.3.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.13
oss2==2.18.1
ossfs==2023.12.0
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pathspec==0.12.1
pendulum==3.0.0
pickle-mixin==1.0.2
pillow==10.2.0
platformdirs==3.11.0
pluggy==1.4.0
portalocker==2.8.2
preshed==3.0.9
prison==0.2.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
pyarrow==10.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
pydot==1.4.2
PyDrive2==1.19.0
pyfarmhash==0.3.2
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
pyjsparser==2.7.1
PyJWT==2.8.0
pymongo==4.6.1
pyOpenSSL==24.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-nvd3==0.15.0
python-slugify==8.0.4
pytz==2023.4
PyYAML==6.0.1
querystring-parser==1.2.4
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==2024.2.0
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
scmrepo==3.1.0
seaborn==0.13.2
semver==3.0.2
sentencepiece==0.2.0
setproctitle==1.3.3
shapely==2.0.2
shortuuid==1.0.11
shtab==1.7.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.51
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
sqltrie==0.11.0
srsly==2.4.8
sshfs==2023.10.0
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-data-validation==1.14.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
tensorflow-serving-api==2.14.1
termcolor==2.4.0
text-unidecode==1.3
tfx-bsl==1.14.0
thinc==8.2.3
threadpoolctl==3.2.0
time-machine==2.13.0
tokenizers==0.15.2
tomlkit==0.12.4
torch==2.0.1
tqdm==4.66.1
transformers==4.38.2
triton==2.0.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uc-micro-py==1.0.3
ujson==5.9.0
unicodecsv==0.14.1
universal_pathlib==0.2.1
uritemplate==3.0.1
urllib3==2.0.7
uvicorn==0.27.1
uvloop==0.19.0
vine==5.1.0
voluptuous==0.14.2
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
webdav4==0.9.8
webencodings==0.5.1
websockets==12.0
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.1.2
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
zstandard==0.22.0
